correcao de rotas e revisoes solicitadas
