+++ front_end/web.py
@@ -5,7 +5,6 @@ sys.path.append('.')
 from flask import Flask, render_template, request, redirect, url_for
 import back_end.controller.ctrl_category as ct_category
 import back_end.controller.ctrl_product as ct_product
-import back_end.controller.ctrl_marketplace as ct_marketplace
 import back_end.controller.ctrl_seller as ct_seller
 import back_end.dao.dao_sql_log as ac_log
 from back_end.models.Marketplace import Marketplace
@@ -13,6 +12,8 @@ from back_end.models.Category import Category
 from back_end.models.product import Product
 from back_end.models.seller import Seller
 from back_end.models.log import Log
+from back_end.controller.ctrl_marketplace import MarketplaceController
+from back_end.controller.ctrl_category import CategoryController
 
 app = Flask(__name__)
 titulo_head = 'Lojinha'
@@ -24,7 +25,7 @@ def cadastro_Marketplace():
     description_market = request.args.get('description')
     if marketplace_add is not None and description_market is not None:
         mkp = Marketplace(marketplace_add, description_market)
-        ct_marketplace.create_marketplace(mkp)
+        MarketplaceController().create(mkp)
         menssagem = f'{mkp.name} cadastrado com sucesso'
     return render_template('create_marketplace.html', menssagem=menssagem, titulo='Cadastro de Marketplaces', titulo_head=titulo_head)
 
@@ -51,9 +52,12 @@ def cadastro_Categoria():
     if category_name is not None:
         category_description = request.args.get('description')
         cat = Category(category_name, category_description)
-        ct_category.create_category(cat)
+        CategoryController.create(cat)
         mensagem = f'{cat.name} cadastrado com sucesso'
     return render_template('create_category.html', menssagem=mensagem, titulo='Cadastro de Categorias', titulo_head=titulo_head)
+
+
+
 
 @app.route('/cadastrar_seller')
 def cadastro_Seller():
